requirements.txt
# 📌 Backend: API, Databas och Autentisering
firebase-admin>=6.0.1  # Firebase Admin SDK för autentisering och Firestore
python-dotenv==1.0.0  # Laddar miljövariabler från .env-fil
requests==2.28.1  # HTTP-bibliotek för API-förfrågningar
flask==2.2.2  # Mikro-webbramverk för API och backend-logik
werkzeug==2.2.2  # WSGI-verktyg för Flask-applikationer
jinja2==3.1.2  # Mallmotor för rendering av HTML-sidor i Flask
itsdangerous==2.1.2  # Säker hantering av dataöverföring i API:er
click==8.1.3  # Kommandoradsgränssnitt för Flask-applikationer
bcrypt==4.0.1  # Hashning av lösenord för säker lagring

# 📌 Röststyrning & Ljudhantering
whisper==1.1.10  # OpenAI:s system för automatisk taligenkänning (ASR)

# 📌 Testning och Debugging
pytest>=7.0.0  # Enhetstestning av backend och funktionalitet
selenium>=4.0.0  # Automatiserad testning av UI i frontend

